Added proper lerping between colors.
I've added both HSV and RGB lerping, as well as cosine interpolation.
I also cleaned up some code, removing `this->` if it doesn't make things clearer.
Removed `_rndColor()` and the option to generate a random seed row as it's not very interesting and doesn't have any effect due to statistical effects at larger scales.

Instead of returning `sf::Color` objects from the palette code it returns a pointer to the color instead if possible. Hopefully this requires fewer allocations, moves, copies, etc.

The HSV/RGB conversion code was borrowed from https://stackoverflow.com/questions/3018313/algorithm-to-convert-rgb-to-hsv-and-hsv-to-rgb-in-range-0-255-for-both
The lerping code was inspired by https://github.com/SFML/SFML/wiki/Source%3A-Color-Gradient